refactor(gates): where a write sits moves to a file of its own
tableownership_test.go crossed the 1000-line test ceiling. The split follows the
seam already in the file: that half asks WHAT a package writes, this one answers
WHERE the write sits, and the three arms of that answer — a method's receiver, a
grouped block's spec, a spec's bound name — are one argument worth reading in
one place.

The new file declares no Test function, so `readGateDeclarations` asks it for no
`//gate:kind` line and the gate inventory keeps one row for this gate rather
than gaining a second. Moving the TEST instead would have added a row, and a
whole-tree manifest is a page every open PR is racing.
